requirements.txt
streamlit>=1.35
pandas>=2.2
pytesseract>=0.3.10
Pillow>=10.0
groq>=0.9
